requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
mlflow>=2.7.0
google-generativeai>=0.5.0
streamlit>=1.28.0
newspaper3k>=0.2.8
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
accelerate>=0.24.0
lxml[html_clean]